gke-devsecops-survival-kit/scripts/requirements.txt
google-cloud-bigquery>=3.11
google-auth>=2.22
